+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/page/AbstractPage.java
@@ -3,7 +3,6 @@ package uk.ac.bbsrc.tgac.miso.webapp.integrationtest.page;
 import static org.openqa.selenium.support.ui.ExpectedConditions.stalenessOf;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
@@ -12,14 +11,10 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedCondition;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.Select;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import uk.ac.bbsrc.tgac.miso.webapp.integrationtest.page.element.AbstractElement;
 
 public abstract class AbstractPage extends AbstractElement {
-
-  private static final Logger log = LoggerFactory.getLogger(AbstractPage.class);
 
   private static final ExpectedCondition<Boolean> pageLoaded = (driver) -> {
     return ((JavascriptExecutor) driver).executeScript("return document.readyState").equals("complete");
@@ -86,30 +81,6 @@ public abstract class AbstractPage extends AbstractElement {
     WebElement selectedButton = buttons.stream().filter(button -> button.isSelected()).findAny().orElse(null);
     if (selectedButton == null) throw new IllegalArgumentException("No buttons are selected for set " + buttons.get(0).getAttribute("id"));
     return selectedButton.getAttribute("value");
-  }
-
-  private static final String MISO_URL = "%smiso/%s";
-  private static final String MISO_STACKTRACE = "uk.ac.bbsrc";
-
-  public static boolean checkForErrors(WebDriver driver, String baseUrl, String urlSlug) {
-    String url = String.format(MISO_URL, baseUrl, urlSlug);
-    try {
-      driver.get(url);
-      // confirm that page contains logo
-      if (driver.findElements(By.id("misologo")).isEmpty())
-        throw new IllegalArgumentException("Page at /miso/" + urlSlug + " is completely empty. Is resource correct?");
-
-      List<WebElement> errors = driver.findElements(By.xpath("//li[contains(text(), '" + MISO_STACKTRACE + "')]"));
-      if (errors.size() > 0) {
-        errors.stream().map(item -> item.getText().trim()).collect(Collectors.toList()).toString();
-        return true;
-      } else {
-        return false;
-      }
-    } catch (Exception e) {
-      log.error(String.format("Exception thrown while checking slug: %s", urlSlug), e);
-      throw e;
-    }
   }
 
   protected WebElement getHtmlElement() {